Add from_pem_spki function to PublicKey impl
This commit adds a function named from_pem_spki to the PublicKey
implementation which accepts a Subject Public Key Info structure in PEM
format.

The motivation for adding this is that I found it useful when one has a
public key in this format, and to be able to create a PublicKey from
it. Perhaps others will find it useful as well.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>